+++ src/main/java/zone/moddev/mc/ironagefurniture/init/ClientItemInitialiser.java
@@ -1,6 +1,8 @@
 package zone.moddev.mc.ironagefurniture.init;
 
 import zone.moddev.mc.ironagefurniture.Ironagefurniture;
+import zone.moddev.mc.ironagefurniture.api.Enumerations.PaddedBenchColour;
+import zone.moddev.mc.ironagefurniture.api.Items.ItemBlockPaddedBench;
 
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
@@ -19,8 +21,16 @@ public final class ClientItemInitialiser {
     public static void registerItemRenders() {
         for (String name : Ironagefurniture.ItemRegistry.keySet()) {
             Item item = Ironagefurniture.ItemRegistry.get(name);
-            Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(item, 0,
-                    new ModelResourceLocation(Ironagefurniture.MODID + ":" + name, "inventory"));
+            ModelResourceLocation model = new ModelResourceLocation(
+                    Ironagefurniture.MODID + ":" + name, "inventory");
+            if (item instanceof ItemBlockPaddedBench) {
+                for (PaddedBenchColour colour : PaddedBenchColour.values()) {
+                    Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(
+                            item, colour.getItemMetadata(), model);
+                }
+            } else {
+                Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(item, 0, model);
+            }
         }
     }
 }